Evaluate given block in the current module context
I'm guessing the original intent of the blocks provided to
MatcherFactory were to mimic blocks provided to RSpec::Matchers.define
calls. In that spirit, evaluate the provided block in the module context
of RSpec's original matcher definition method so that any methods
defined within the block are defined on the module/class rather than in
a main/global context.

This alleviates warning messages about the traverse method already being
defined in the context and getting redefined.